[SGH-38] Added fnc for removing associations from joined tables

+++ src/main/java/edu/codespring/sportgh/service/CategoryServiceImpl.java
@@ -23,8 +23,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long categoryId) {
-        categoryRepository.deleteById(categoryId);
-        log.info("Category with ID {} deleted successfully.", categoryId);
+        Category category = categoryRepository.findById(categoryId).orElse(null);
+        if (category != null) {
+            category.deleteAssociations();
+            categoryRepository.deleteById(categoryId);
+            log.info("Category with ID {} deleted successfully.", categoryId);
+        }
     }
 
     @Override
